pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "data-engineering-pipeline"
version = "0.1.0"
description = "Scalable ETL and data workflow pipelines"
authors = ["leeroo-coder <leerooadmin@leeroo.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"
structlog = "^23.0.0"
prefect = "^2.0.0"
boto3 = "^1.26.0"
neo4j = "^5.0.0"
great-expectations = "^0.17.0"
networkx = "^3.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
click = "^8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
ruff = "^0.0.280"
mypy = "^1.0.0"
pre-commit = "^3.0.0"

[tool.poetry.group.spark.dependencies]
pyspark = "^3.4.0"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "RUF"]
ignore = ["S101", "S608", "PLR0913", "PLR0912", "PLR0915"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.poetry.scripts]
etl-pipeline = "src.main:main"